added error handling and maybe broke some stuff

+++ src/main/kotlin/ua/leonidius/garage/graphql/car_detail/CarDetailMutation.kt
@@ -14,17 +14,21 @@ class CarDetailMutation: GraphQLMutationResolver {
     private lateinit var detailService: CarDetailServiceFacade
 
     fun createDetail(name: String, manufacturer: String, description: String,
-                     type: String, price: Float): CarDetailDto {
+                     type: String, price: Float): CarDetailDto? {
+        if (price < 0) return null
         return detailService.addCarDetail(name, manufacturer, description, price, type)
     }
 
-    fun deleteDetail(id: Int): Int {
+    fun deleteDetail(id: Int): Int? {
+        if (id < 0) return null
         detailService.deleteDetail(id)
         return id
     }
 
     fun updateDetail(id: Int, name: String?, manufacturer: String?, description: String?,
                      type: String?, price: Float?): Optional<CarDetailDto> {
+        if (id < 0 || (price != null && price < 0)) return Optional.empty()
+
         val detail = detailService.updateLocalDetail(id, name, manufacturer, description, type, price)
         if (detail == null) return Optional.empty()
         else return Optional.of(detail)
